requirements.txt
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
gologin==2025.8.1
google-auth==2.42.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
h11==0.16.0
httplib2==0.31.0
idna==3.11
oauth2client==4.1.3
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
psutil==7.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.37.0
sentry-sdk==2.42.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
